+++ packages/web3-core/CHANGELOG.md
@@ -193,6 +193,10 @@ Documentation:
 
 ## [4.3.1]
 
+### Fixed
+
+-   Fix `Web3Config` to properly update within other web3 packages when `setConfig` is used  (#6555)
+
 ### Added
 
 -   Added `isMetaMaskProvider` function to check if provider is metamask (#6534)